SEM3/SPA/Assignment1/PulsarRealtime/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
pulsar-client==3.6.1
eventlet==0.33.3
python-dotenv==1.0.0